# openapi.yaml
openapi: 3.0.0
info:
  title: 'Q Laravel Skeleton API'
  version: '0.9'
paths:
...
  '/api/v1/authors/{author}':
    get:
      tags:
        - Authors
      summary: 'Get author'
      operationId: getAuthor
      parameters:
        -
          name: author
          in: path
          description: 'Author id'
          required: true
          example: 1
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                properties:
                  author: { $ref: '#/components/schemas/AuthorResource' }
                type: object
        '400':
          description: 'Bad Request'
        '401':
          description: Unauthorized
        '404':
          description: 'Not found'
      security:
        -
          sanctum: []
    put:
      tags:
        - Authors
      summary: 'Update existing author'
      operationId: updateAuthor
      parameters:
        -
          name: author
          in: path
          description: 'Author id'
          required: true
          example: 1
      requestBody:
        description: 'Update existing author'
        required: true
        content:
          application/json:
            schema:
              required:
                - first_name
                - last_name
              properties:
                first_name:
                  type: string
                  example: John
                last_name:
                  type: string
                  example: Doe
                birthday:
                  type: string
                  example: '1980-01-23'
                biography:
                  type: string
                  example: 'John Doe was amazing author who ...'
                place_of_birth:
                  type: string
                  example: London
              type: object
      responses:
        '200':
          description: 'Author updated successfully'
          content:
            application/json:
              schema:
                properties:
                  author: { $ref: '#/components/schemas/AuthorResource' }
                type: object
        '400':
          description: 'Bad Request'
        '401':
          description: Unauthorized
        '404':
          description: 'Not found'
        '422':
          description: 'Validation failed'
      security:
        -
          sanctum: []
    delete:
      tags:
        - Authors
      summary: 'Delete author'
      operationId: deleteAuthor
      parameters:
        -
          name: author
          in: path
          description: 'Author id'
          required: true
          example: 1
      responses:
        '204':
          description: Deleted
          content:
            application/json:
              schema:
                properties:
                  message: { type: string, example: 'Author deleted.' }
                type: object
        '400':
          description: 'Bad Request'
        '401':
          description: Unauthorized
        '404':
          description: 'Not found'
      security:
        -
          sanctum: []
  /api/v1/books:
    get:
      tags:
        - Books
      summary: 'Get list of books'
      operationId: getBooks
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                properties:
                  book: { $ref: '#/components/schemas/BookResource' }
                type: object
        '400':
          description: 'Bad Request'
        '401':
          description: Unauthorized
      security:
        -
          sanctum: []
    post:
      tags:
        - Books
      summary: 'Create new book'
      operationId: createBook
      requestBody:
        description: 'Create new book'
        required: true
        content:
          application/json:
            schema:
              required:
                - isbn
                - title
              properties:
                isbn:
                  type: string
                  example: 978-92-95055-02-5
                title:
                  type: string
                  example: 'How to tie your shoes'
                release_date:
                  type: string
                  example: '1999-11-25'
                format:
                  type: string
                  example: PDF
                pages:
                  type: integer
                  example: '218'
                publisher_id:
                  type: integer
                  example: '1'
              type: object
      responses:
        '201':
          description: 'Book created successfully'
          content:
            application/json:
              schema:
                properties:
                  book: { $ref: '#/components/schemas/BookResource' }
                type: object
        '400':
          description: 'Bad Request'
        '401':
          description: Unauthorized
        '422':
          description: 'Validation failed'
      security:
        -
          sanctum: []
  '/api/v1/books/{book}':
    get:
      tags:
        - Books
      summary: 'Get book'
      operationId: getBook
      parameters:
        -
          name: book
          in: path
          description: 'Book id'
          required: true
          example: 1
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                properties:
                  book: { $ref: '#/components/schemas/BookResource' }
                type: object
        '400':
          description: 'Bad Request'
        '401':
          description: Unauthorized
        '404':
          description: 'Not found'
      security:
        -
          sanctum: []
    put:
      tags:
        - Books
      summary: 'Update existing book'
      operationId: updateBook
      parameters:
        -
          name: book
          in: path
          description: 'Book id'
          required: true
          example: 1
      requestBody:
        description: 'Update existing book'
        required: true
        content:
          application/json:
            schema:
              required:
                - isbn
                - title
              properties:
                isbn:
                  type: string
                  example: 978-92-95055-02-5
                title:
                  type: string
                  example: 'How to tie your shoes'
                release_date:
                  type: string
                  example: '1999-11-25'
                format:
                  type: string
                  example: PDF
                pages:
                  type: integer
                  example: '218'
                publisher_id:
                  type: integer
                  example: '1'
              type: object
      responses:
        '200':
          description: 'Book updated successfully'
          content:
            application/json:
              schema:
                properties:
                  book: { $ref: '#/components/schemas/BookResource' }
                type: object
        '400':
          description: 'Bad Request'
        '401':
          description: Unauthorized
        '404':
          description: 'Not found'
        '422':
          description: 'Validation failed'
      security:
        -
          sanctum: []
    delete:
      tags:
        - Books
      summary: 'Delete book'
      operationId: deleteBook
      parameters:
        -
          name: book
          in: path
          description: 'Book id'
          required: true
          example: 1
      responses:
        '204':
          description: Deleted
          content:
            application/json:
              schema:
                properties:
                  message: { type: string, example: 'Book deleted.' }
                type: object
        '400':
          description: 'Bad Request'
        '401':
          description: Unauthorized
        '404':
          description: 'Not found'
      security:
        -
          sanctum: []
  /api/v1/publishers:
    get:
      tags:
        - Publishers
      summary: 'Get list of publishers'
      operationId: getPublishers
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                properties:
                  publisher: { $ref: '#/components/schemas/PublisherResource' }
                type: object
        '400':
          description: 'Bad Request'
        '401':
          description: Unauthorized
      security:
        -
          sanctum: []
    post:
      tags:
        - Publishers
      summary: 'Create new publisher'
      operationId: createPublisher
      requestBody:
        description: 'Create new publisher'
        required: true
        content:
          application/json:
            schema:
              required:
                - name
              properties:
                name:
                  type: string
                  example: 'Q Publishing'
                address:
                  type: string
                  example: 'Some address 33'
              type: object
      responses:
        '201':
          description: 'Publisher created successfully'
          content:
            application/json:
              schema:
                properties:
                  publisher: { $ref: '#/components/schemas/PublisherResource' }
                type: object
        '400':
          description: 'Bad Request'
        '401':
          description: Unauthorized
        '422':
          description: 'Validation failed'
      security:
        -
          sanctum: []
  '/api/v1/publishers/{publisher}':
    get:
      tags:
        - Publishers
      summary: 'Get publisher'
      operationId: getPublisher
      parameters:
        -
          name: publisher
          in: path
          description: 'Publisher id'
          required: true
          example: 1
      responses:
        '200':
          description: success
          content:
            application/json:
              schema:
                properties:
                  publisher: { $ref: '#/components/schemas/PublisherResource' }
                type: object
        '400':
          description: 'Bad Request'
        '401':
          description: Unauthorized
        '404':
          description: 'Not found'
      security:
        -
          sanctum: []
    put:
      tags:
        - Publishers
      summary: 'Update existing publisher'
      operationId: updatePublisher
      parameters:
        -
          name: publisher
          in: path
          description: 'Publisher id'
          required: true
          example: 1
      requestBody:
        description: 'Update existing publisher'
        required: true
        content:
          application/json:
            schema:
              required:
                - name
              properties:
                name:
                  type: string
                  example: 'Q Publishing'
                address:
                  type: string
                  example: 'Some address 33'
              type: object
      responses:
        '200':
          description: 'Publisher updated successfully'
          content:
            application/json:
              schema:
                properties:
                  publisher: { $ref: '#/components/schemas/PublisherResource' }
                type: object
        '400':
          description: 'Bad Request'
        '401':
          description: Unauthorized
        '404':
          description: 'Not found'
        '422':
          description: 'Validation failed'
      security:
        -
          sanctum: []
    delete:
      tags:
        - Publishers
      summary: 'Delete publisher'
      operationId: deletePublisher
      parameters:
        -
          name: publisher
          in: path
          description: 'Publisher id'
          required: true
          example: 1
      responses:
        '204':
          description: Deleted
          content:
            application/json:
              schema:
                properties:
                  message: { type: string, example: 'Publisher deleted.' }
                type: object
        '400':
          description: 'Bad Request'
        '401':
          description: Unauthorized
        '404':
          description: 'Not found'
      security:
        -
          sanctum: []
...
components:
  schemas:
    AuthorResource:
      description: 'Class AuthorResource'
      properties:
        id:
          type: integer
          default: 1
        first_name:
          type: string
          default: John
        last_name:
          type: string
          default: Doe
        birthday:
          type: string
          default: '1980-01-23'
        biography:
          type: string
          default: 'John Doe was amazing author who ...'
        place_of_birth:
          type: string
          default: London
      type: object
    BookResource:
      description: 'Class BookResource'
      properties:
        id:
          type: integer
          default: 1
        isbn:
          type: string
          default: 978-92-95055-02-5
        title:
          type: string
          default: 'How to tie your shoes'
        release_date:
          type: string
          default: '1999-25-11'
        format:
          type: string
          default: PDF
        pages:
          type: integer
          default: '218'
        publisher:
          $ref: '#/components/schemas/PublisherResource'
      type: object
    PublisherResource:
      description: 'Class PublisherResource'
      properties:
        id:
          type: integer
          default: 1
        name:
          type: string
          default: 'Q Books Ltd'
        address:
          type: string
          default: '5th Avenue, LA'
        books:
          $ref: '#/components/schemas/BookResource'
      type: object
...